+++ db/migrate/20251020011346_rename_task_run_args_to_run_with.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class RenameTaskRunArgsToRunWith < ActiveRecord::Migration[7.1]
+  def change
+    rename_column :raif_tasks, :task_run_args, :run_with
+  end
+end