+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add support for Azure OpenAI's API key authentication header style. [#263](https://github.com/CultivateLabs/raif/pull/263)
 - Add admin page for viewing current configuration. [#264](https://github.com/CultivateLabs/raif/pull/264)
 - Make it so developer can define their own, custom agent final answer tool (e.g. if you want a custom schema for the response). [##266](https://github.com/CultivateLabs/raif/pull/#266)
+- `task_run_arg` has been renamed to `run_with` and is now supported by `Raif::Agent`. [#269](https://github.com/CultivateLabs/raif/pull/269)
 
 ## v1.3